refactor: extract log function to src/logger.js (roadmap step 2)
Lift log() and the levels map out of main.js into a dedicated logger
module. main.js retains a transitional re-export with a TODO marker
(removed in step 3) so nothing breaks mid-migration. renderer.js now
imports log directly from logger.js, killing one import cycle.

tests/renderer.test.js mock setup rewired: log is now mocked via
jest.mock('../src/logger.js', ...) instead of through the main.js mock.

No behavior change. Test baseline: 13/13 passing before and after.
Smoke tested in Chrome: LCD renders, log textarea updates, no console errors.

Per docs/refactor/04-roadmap.md step 2.
